Sales_DE_Project/requirements.txt
Flask==3.0.0
google-cloud-storage==2.14.0
Werkzeug==3.0.1
